Pattern: fix numerator calculation

+++ src/core/Basics/Pattern.cpp
@@ -285,7 +285,8 @@ void Pattern::saveTo( XMLNode& node, int nInstrumentId, const QString& sType,
 }
 
 float Pattern::numerator() const {
-	return m_nLength  * m_nDenominator / ( 4 * H2Core::nTicksPerQuarter );
+	return static_cast<float>(m_nLength) * static_cast<float>(m_nDenominator) /
+		( 4 * static_cast<float>(H2Core::nTicksPerQuarter) );
 }
 
 std::shared_ptr<Note> Pattern::findNote( int nPosition, int nInstrumentId,
